requirements.txt
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-SQLALchemy==3.0.3
Flask-JWT-Extended==4.5.2
Flask-RESTful==0.3.10
Jinja2==3.1.2
pyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.2.2
bcrypt==4.0.1
flask-marshmallow==0.15.0
mysqlclient==2.2.0